fix: corrigindo merge

+++ ms-nutria-mongodb/src/main/java/org/example/msnutriamongodb/repository/ChatRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ChatRepository extends MongoRepository<Chat, Long> {}
+public interface ChatRepository extends MongoRepository<Chat, Integer> {}
